Added Role Based Access

+++ src/account/service/UserService.java
@@ -1,16 +1,21 @@
 package account.service;
 
-import account.model.NewPasswordDto;
-import account.model.User;
+import account.model.*;
+import account.repository.RoleRepository;
 import account.repository.UserRepository;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import javax.transaction.Transactional;
 import java.security.Principal;
+import java.util.List;
+import java.util.Set;
+
+import static account.model.Operation.GRANT;
+import static account.model.Operation.REMOVE;
+import static account.model.Role.*;
 
 /**
  * @author adnan
@@ -20,11 +25,51 @@ import java.security.Principal;
 public class UserService {
 
     UserRepository userRepository;
+    RoleRepository roleRepository;
     PasswordEncoder encoder;
 
-    public UserService(UserRepository userRepository, PasswordEncoder encoder) {
+    public UserService(UserRepository userRepository, RoleRepository roleRepository, PasswordEncoder encoder) {
         this.userRepository = userRepository;
+        this.roleRepository = roleRepository;
         this.encoder = encoder;
+    }
+
+    public User changeRole(RoleChangeDto roleChangeDto) {
+        String email = roleChangeDto.getUser();
+        String roleName = ROLE_PREFIX + roleChangeDto.getRole();
+        Operation operation = roleChangeDto.getOperation();
+
+        User user = findByEmail(email);
+        Role role = roleRepository.findTopByName(roleName)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Role not found!"));
+
+        if (operation == GRANT) {
+            if ((user.isAdmin() && !role.getName().equals(ROLE_ADMINISTRATOR))
+                    || (!user.isAdmin() && role.getName().equals(ROLE_ADMINISTRATOR))) {
+
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "The user cannot combine administrative and business roles!");
+            }
+
+            user.getRoles().add(role);
+            user = saveOrUpdate(user);
+        } else if (operation == REMOVE) {
+            if (!user.getRoles().contains(role)) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "The user does not have a role!");
+            }
+
+            if (user.isAdmin()) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Can't remove ADMINISTRATOR role!");
+            }
+
+            if (user.getRoles().size() == 1) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "The user must have at least one role!");
+            }
+
+            user.getRoles().remove(role);
+            user = saveOrUpdate(user);
+        }
+
+        return user;
     }
 
     public void changePassword(NewPasswordDto newPasswordDto, Principal principal) {
@@ -38,6 +83,10 @@ public class UserService {
         saveOrUpdate(user);
     }
 
+    public List<User> findAll() {
+        return userRepository.findAllByOrderByIdAsc();
+    }
+
     public User findById(Long id) {
         return userRepository.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
@@ -45,7 +94,7 @@ public class UserService {
 
     public User findByEmail(String email) {
         return userRepository.findByEmail(email.toLowerCase())
-                .orElseThrow(() -> new UsernameNotFoundException("No such user with this email!"));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found!"));
     }
 
     @Transactional
@@ -53,13 +102,22 @@ public class UserService {
         if (user.isNew()) {
             user.setEmail(user.getEmail().toLowerCase());
             user.setPassword(encoder.encode(user.getPassword()));
+            user.setRoles(userRepository.count() == 0
+                    ? Set.of(roleRepository.findTopByName(ROLE_ADMINISTRATOR).orElse(new Role(ROLE_ADMINISTRATOR)))
+                    : Set.of(roleRepository.findTopByName(ROLE_USER).orElse(new Role(ROLE_USER))));
         }
 
         return userRepository.save(user);
     }
 
     @Transactional
-    public void delete(User user) {
+    public void delete(String email) {
+        User user = findByEmail(email);
+
+        if (user.isAdmin()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Can't remove ADMINISTRATOR role!");
+        }
+
         userRepository.delete(user);
     }
 }
